refactor(config): simplify app config wiring
