+++ packages/core/src/types/auth.types.ts
@@ -55,11 +55,3 @@ export const GoogleConnectErrorResponseSchema = ApiErrorResponseSchema.extend({
 export type GoogleConnectErrorResponse = z.infer<
   typeof GoogleConnectErrorResponseSchema
 >;
-
-export const AppConfigSchema = z.object({
-  google: z.object({
-    isConfigured: z.boolean(),
-  }),
-});
-
-export type AppConfig = z.infer<typeof AppConfigSchema>;
