+++ packages/core/src/types/config.types.ts
@@ -1,0 +1,9 @@
+import { z } from "zod";
+
+export const AppConfigSchema = z.object({
+  google: z.object({
+    isConfigured: z.boolean(),
+  }),
+});
+
+export type AppConfig = z.infer<typeof AppConfigSchema>;
